+++ src/rules/util.ts
@@ -21,6 +21,8 @@ export const createRule = ESLintUtils.RuleCreator(
     `https://github.com/nebrius/eslint-plugin-fast-import/tree/main/src/rules/${name}/README.md`
 );
 
+// TODO: This should also be project dependent so that we don't call update
+// listeners for files that are part of a different project
 const updateListeners = new Set<() => void>();
 export function registerUpdateListener(cb: () => void) {
   updateListeners.add(cb);